refactor home view
