+++ lib/controllers/HomeController/HomeController.dart
@@ -5,8 +5,8 @@ import 'package:Videotheque/services/FirestoreQueries.dart';
 import 'package:Videotheque/Globals.dart';
 import 'package:Videotheque/utils/Singletons.dart';
 import 'package:Videotheque/utils/CustomChangeNotifier.dart';
-import 'package:Videotheque/views/home_view/CarrouselView.dart';
-import 'package:Videotheque/views/home_view/components/InfoComponent.dart';
+import 'package:Videotheque/views/HomeView/CarrouselView.dart';
+import 'package:Videotheque/views/HomeView/components/InfoComponent.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:in_app_review/in_app_review.dart';
