+++ lib/views/HomeView/components/AppBarComponent.dart
