+++ lib/views/HomeView/components/InfoComponent.dart
